// device-activation-portal-angular/src/app/shared/model/loginResponse.ts
/**
 * Device Activation API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface LoginResponse { 
    succes?: boolean;
    errorMessage?: string;
    token?: string;
    rules?: Array<string>;
}